requirements.txt
langchain==0.1.5
langchain-openai==0.0.5
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0